% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_block.R
\name{as.data.frame.RandomBlock}
\alias{as.data.frame.RandomBlock}
\title{Convert RandomBlock Arms to Data Frame}
\usage{
\method{as.data.frame}{RandomBlock}(x, ...)
}
\arguments{
\item{x}{A \code{RandomBlock} reference class object.}

\item{...}{Additional arguments passed to the conversion function.}
}
\value{
A data frame containing all block arms from the block.
}
\description{
Converts the list of \code{RandomBlockArm} objects stored in a \code{RandomBlock}
instance into a data frame by applying \code{.treatmentListToDataFrame} to the block arms.
}
\seealso{
\code{\link[=RandomBlock]{RandomBlock}}
}
