% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_data_base.R
\name{as.data.frame.RandomDataBase}
\alias{as.data.frame.RandomDataBase}
\title{Convert RandomDataBase Subjects to Data Frame}
\usage{
\method{as.data.frame}{RandomDataBase}(x, ...)
}
\arguments{
\item{x}{A \code{RandomDataBase} reference class object.}

\item{...}{Additional arguments passed to \code{as.data.frame}.}
}
\value{
A data frame containing all subjects from the database.
}
\description{
Converts the list of \code{RandomSubject} objects stored in a \code{RandomDataBase} instance into a data frame by applying \code{as.data.frame} to each subject and binding the results by row.
}
\seealso{
\code{\link[=RandomDataBase]{RandomDataBase}}
}
