% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_system_state.R
\name{as.data.frame.RandomSystemState}
\alias{as.data.frame.RandomSystemState}
\title{Convert RandomSystemState to Data Frame}
\usage{
\method{as.data.frame}{RandomSystemState}(x, ...)
}
\arguments{
\item{x}{A \code{RandomSystemState} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame with columns for overall and block-wise treatment arm filling levels.
}
\description{
Converts a \code{RandomSystemState} object into a data frame containing overall and block-wise treatment arm filling levels.
}
\seealso{
\code{\link[=RandomSystemState]{RandomSystemState}}
}
\keyword{internal}
