% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_random_block_size.R
\name{getBlockSizes}
\alias{getBlockSizes}
\title{Get Block Sizes for Treatment Arms}
\usage{
getBlockSizes(treatmentArmIds, blockSizes, allocationFraction = NA_real_)
}
\arguments{
\item{treatmentArmIds}{Character vector of treatment arm identifiers.}

\item{blockSizes}{Integer vector specifying the total block sizes to use.
If more than one block size is provided, variable block sizes will be used.}

\item{allocationFraction}{Numeric vector specifying the allocation
fraction for each treatment arm. Defaults to equal allocation
if \code{NA}.}
}
\value{
A list of block allocations, each represented as a named
list of counts per treatment arm.
}
\description{
Generates a list of block sizes per treatment arm for specified treatment arms and
total block sizes. Each block size per treatment arm is computed according to the provided
allocation fraction or defaults to equal allocation if not specified.
}
\examples{
# variable block sizes with 1:1 allocation ratio
getBlockSizes(c("A", "B"), c(4, 6))

# fixed block size with 2:1 allocation ratio
getBlockSizes(c("A", "B"), 9, c(2/3, 1/3))

}
