% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_allocation_value_service.R
\name{getRandomAllocationValueService}
\alias{getRandomAllocationValueService}
\title{Create a New RandomAllocationValueService Instance}
\usage{
getRandomAllocationValueService()
}
\value{
A \code{RandomAllocationValueService} reference class object.
}
\description{
Constructs and returns a new \code{RandomAllocationValueService} reference
class object, which manages the generation and retrieval of random
allocation values for randomization procedures.
}
\seealso{
\code{\link[=RandomAllocationValueService]{RandomAllocationValueService}}
}
