% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_configuration.R
\name{getRandomConfiguration}
\alias{getRandomConfiguration}
\title{Create a New RandomConfiguration Instance}
\usage{
getRandomConfiguration(
  ...,
  randomProject,
  treatmentArmIds,
  seed = NA_integer_,
  ravBufferMinimumSize = 1000L,
  ravBufferMaximumSize = 10000L
)
}
\arguments{
\item{...}{Additional arguments passed to the constructor.}

\item{randomProject}{\code{RandomProject} object associated with this configuration.}

\item{treatmentArmIds}{Character vector of treatment arm IDs.}

\item{seed}{Integer random seed used for reproducibility. Defaults to \code{NA_integer_}.}

\item{ravBufferMinimumSize}{Integer specifying the minimum buffer size for allocation values. Defaults to \code{1000L}.}

\item{ravBufferMaximumSize}{Integer specifying the maximum buffer size for allocation values. Defaults to \code{10000L}.}
}
\value{
A \code{RandomConfiguration} reference class object.
}
\description{
Constructs and returns a new \code{RandomConfiguration} reference class object,
initializing all fields and assigning a unique ID. Allows specification of
treatment arms, seed, buffer sizes, and associated project.
}
\seealso{
\code{\link[=RandomConfiguration]{RandomConfiguration}}
}
