% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_data_base.R
\name{getRandomDataBase}
\alias{getRandomDataBase}
\title{Create a New RandomDataBase Instance}
\usage{
getRandomDataBase()
}
\value{
A \code{RandomDataBase} reference class object.
}
\description{
Constructs and returns a new \code{RandomDataBase} reference class object,
initializing all fields and assigning a unique ID.
}
\seealso{
\code{\link[=RandomDataBase]{RandomDataBase}}

\code{\link[=as.data.frame.RandomDataBase]{as.data.frame()}}
}
