% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_project.R
\name{getRandomProject}
\alias{getRandomProject}
\title{Create a RandomProject Instance}
\usage{
getRandomProject(name)
}
\arguments{
\item{name}{Character string specifying the project name.}
}
\value{
A \code{RandomProject} reference class object.
}
\description{
Instantiates a new \code{RandomProject} object with the specified project name.
}
\seealso{
\code{\link[=RandomProject]{RandomProject}}
}
