% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_allocation_value_service.R
\name{plot.RandomAllocationValueService}
\alias{plot.RandomAllocationValueService}
\title{Plot Distribution of Random Allocation Values}
\usage{
\method{plot}{RandomAllocationValueService}(x, ..., usedValuesOnly = TRUE)
}
\arguments{
\item{x}{A \code{RandomAllocationValueService} object.}

\item{...}{Additional arguments passed to the \code{hist} function.}

\item{usedValuesOnly}{Logical; if \code{TRUE}, only values up to the current index are plotted.}
}
\value{
A histogram plot of the random allocation values.
}
\description{
Visualizes the distribution of random allocation values managed by
a \code{RandomAllocationValueService} object. Optionally restricts the plot
to only used values and performs a chi-squared test for uniformity.
}
