% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_data_base.R
\docType{class}
\name{RandomDataBase-class}
\alias{RandomDataBase-class}
\alias{RandomDataBase}
\title{RandomDataBase Reference Class}
\description{
Manages randomization data, including projects, configurations,
allocation values, subjects, and results. Provides methods for
validation, persistence, retrieval, and display of randomization objects.
}
\section{Fields}{

\describe{
\item{\code{uniqueId}}{Character string uniquely identifying the database instance.}

\item{\code{creationDate}}{POSIXct timestamp of database creation.}

\item{\code{randomProjects}}{List of \code{RandomProject} objects.}

\item{\code{randomConfigurations}}{List of \code{RandomConfiguration} objects.}

\item{\code{randomAllocationValues}}{List of \code{RandomAllocationValue} objects.}

\item{\code{randomSubjects}}{List of \code{RandomSubject} objects.}

\item{\code{randomResults}}{List of \code{RandomResult} objects.}
}}


\section{Methods}{

\describe{
\item{initialize(..., creationDate)}{Initializes a new \code{RandomDataBase} instance, assigns a unique ID, and sets up empty lists for all fields.}
\item{validateRandomProject(randomProject)}{Checks if a project exists in the database; stops with an error if not.}
\item{show()}{Displays a summary of the database and its contents.}
\item{toString()}{Returns a string representation of the database and its statistics.}
\item{persist(obj)}{Persists an object to the appropriate list based on its class.}
\item{getRandomProjectUniqueIds(objects)}{Returns a character vector of unique IDs for the given objects' projects.}
\item{getLastSubject(randomProject)}{Retrieves the last subject for a given project.}
\item{getLastRandomConfiguration(randomProject)}{Retrieves the last configuration for a given project.}
\item{createNewSubjectRandomNumber(randomProject)}{Generates the next subject random number for a given project.}
}
}

\seealso{
\code{\link[=getRandomDataBase]{getRandomDataBase()}}

\code{\link[=as.data.frame.RandomDataBase]{as.data.frame()}}
}
\keyword{internal}
