% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{randomforge}
\alias{randomforge-package}
\alias{randomforge}
\title{randomforge Package}
\description{
Randomforge provides an extensible and transparent framework for
implementing, evaluating, and applying randomization methods for
clinical trials.
}
\details{
The package aims to support a broad range of classical,
covariate-adaptive, and response-adaptive techniques while
enabling reproducibility, auditability, and methodological clarity.
Built as part of the open RandomForge initiative ("Innovating the
Future of Randomization"), the package encourages community
collaboration, modular extensions, and contributions from academia,
industry, and clinical researchers.

The architecture is designed for integration into larger ecosystems,
including Shiny applications, APIs, and cloud-based workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://randomforge.org}
  \item \url{https://github.com/rpact-com/randomforge}
  \item Report bugs at \url{https://github.com/rpact-com/randomforge/issues}
}

}
\author{
Friedrich Pahlke
}
