% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_method_range_set_pbr.R
\docType{class}
\name{RandomMethodRangeSetPBR-class}
\alias{RandomMethodRangeSetPBR-class}
\alias{RandomMethodRangeSetPBR}
\title{RandomMethodRangeSetPBR Reference Class}
\description{
Extends \code{RandomMethodRangeSet} to support probability block randomization, initializing ranges based on a \code{RandomBlock} object.
}
\section{Fields}{

\describe{
\item{\code{Inherits}}{all fields from \code{RandomMethodRangeSet}.}
}}


\section{Methods}{

\describe{
\item{initialize(block, ...)}{Initializes a new \code{RandomMethodRangeSetPBR} instance using a \code{RandomBlock}. Validates the block and sets up ranges according to its probabilities.}
}
}

\keyword{internal}
