% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_project.R
\docType{class}
\name{RandomProject-class}
\alias{RandomProject-class}
\alias{RandomProject}
\title{RandomProject Reference Class}
\description{
Represents a randomization project, including a unique identifier, project name, and creation date.
}
\section{Fields}{

\describe{
\item{\code{uniqueId}}{Character string uniquely identifying the project.}

\item{\code{name}}{Character string specifying the project name.}

\item{\code{creationDate}}{POSIXct timestamp of project creation.}
}}


\section{Methods}{

\describe{
\item{initialize(..., creationDate = Sys.time())}{Initializes a new \code{RandomProject} instance and assigns a unique ID.}
\item{show(prefix = "")}{Prints a summary of the project.}
\item{toString(prefix = "")}{Returns a string representation of the project.}
}
}

\seealso{
\code{\link[=getRandomProject]{getRandomProject()}}
}
\keyword{internal}
