% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rpact}
\alias{randomforge}
\alias{randomforge-package}
\alias{rpact}
\title{randomforge: Production Grade Randomization}
\description{
Production grade randomization of clinical trials
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.evidat.com}
  \item Report bugs at \url{https://bugreport.evidat.com}
}

}
\author{
Friedrich Pahlke
}
