% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_random_block.R
\docType{class}
\name{RandomBlock-class}
\alias{RandomBlock-class}
\alias{RandomBlock}
\title{Random Block}
\arguments{
\item{treatmentArmIds}{the treatment arm id's.}

\item{maximumBlockSize}{the maximum block length <code>B</code> for each treatment arm.}

\item{factorLevels}{the optional factors and factor levels.}
}
\description{
Random Block
}
\section{Fields}{

\describe{
\item{\code{blockArms}}{The map containing the block arms, 
format: [treatment arm id, RandomBlockArm]}

\item{\code{factorLevels}}{The strata and strata levels formatted as character list
where the key is the factor name or id and the value is the specific level (name or id).}
}}


